+++ src/modules/tools/draw/store/actions/actionsDownload.js
@@ -33,10 +33,10 @@ async function convertFeatures ({state, dispatch}, format) {
  *
  * @returns {void}
  */
-function fileDownloaded ({commit}) {
+function fileDownloaded ({state, commit}) {
     commit("setDownloadEnabled");
     commit("setDownloadFileName", "");
-    commit("setDownloadSelectedFormat", "");
+    commit("setDownloadSelectedFormat", state.download.preSelectedFormat);
 }
 
 /**
@@ -146,11 +146,10 @@ function setDownloadFileName ({state, commit, dispatch}, {currentTarget}) {
  * If the user entered a name for a file and has chosen a format for the features, the download button is enabled.
  *
  * @param {Event} event Event fired by selecting a different element.
- * @param {HTMLSelectElement} event.currentTarget The HTML select element for the file format.
+ * @param {String} value The selected option value from dropdown or pre selected value.
  * @returns {void}
  */
-async function setDownloadSelectedFormat ({state, commit, dispatch}, {currentTarget}) {
-    const {value} = currentTarget;
+async function setDownloadSelectedFormat ({state, commit, dispatch}, value) {
 
     commit("setDownloadSelectedFormat", value);
     if (state.layer.getSource().getFeatures().length > 0) {