+++ src/modules/tools/indexTools.js
@@ -10,6 +10,7 @@ import Contact from "./contact/store/indexContact";
 import Draw from "./draw/store/indexDraw";
 import FileImport from "./fileImport/store/indexFileImport";
 import Gfi from "./gfi/store/indexGfi";
+import SaveSelection from "./saveSelection/store/indexSaveSelection";
 import ScaleSwitcher from "./scaleSwitcher/store/indexScaleSwitcher";
 import SupplyCoord from "./supplyCoord/store/indexSupplyCoord";
 import Measure from "./measure/store/indexMeasure";
@@ -25,6 +26,7 @@ export default {
         Draw,
         FileImport,
         Gfi,
+        SaveSelection,
         ScaleSwitcher,
         SupplyCoord,
         Measure