+++ src/modules/tools/saveSelection/store/indexSaveSelection.js
@@ -1,0 +1,12 @@
+import actions from "./actionsSaveSelection";
+import getters from "./gettersSaveSelection";
+import mutations from "./mutationsSaveSelection";
+import state from "./stateSaveSelection";
+
+export default {
+    namespaced: true,
+    state,
+    actions,
+    mutations,
+    getters
+};