+++ CHANGELOG.md
@@ -22,12 +22,14 @@
 
 ### Deprecated
 - Deprecated the parameter 'serviceID' for the contact tool; 'serviceId' should be used instead.
+- Migrated the SaveSelection Tool from Backbone.js to Vue.js. It also no longer has a Radio channel.
+- copyToClipboard is now a reusable action and can no longer be triggered through the Radio.
+- Deprecated the parameter `simpleMap` in the config.js and moved its configuration to the `saveSelection` tool in the config.json.
 
 ### Removed
 
 ### Fixed
 - If the Save selection tool is not configured, the button for it is no longer displayed in the topic tree.
-
 
 ---